+++ modules/frontend/metrics_query_range_handler_test.go
@@ -1,0 +1,174 @@
+package frontend
+
+import (
+	"net/http/httptest"
+	"testing"
+	"time"
+
+	"github.com/gogo/protobuf/jsonpb"
+	"github.com/gogo/protobuf/proto"
+	"github.com/grafana/dskit/user"
+	"github.com/grafana/tempo/pkg/api"
+	"github.com/grafana/tempo/pkg/tempopb"
+	"github.com/stretchr/testify/require"
+)
+
+func TestQueryRangeHandlerSucceeds(t *testing.T) {
+	resp := &tempopb.QueryRangeResponse{
+		Metrics: &tempopb.SearchMetrics{
+			InspectedTraces: 1,
+			InspectedBytes:  1,
+		},
+		Series: []*tempopb.TimeSeries{
+			{
+				PromLabels: "foo",
+				Samples: []tempopb.Sample{
+					{
+						TimestampMs: 2,
+						Value:       2,
+					},
+					{
+						TimestampMs: 1,
+						Value:       1,
+					},
+				},
+			},
+		},
+	}
+
+	f := frontendWithSettings(t, &mockRoundTripper{
+		responseFn: func() proto.Message {
+			return resp
+		},
+	}, nil, nil, nil)
+	tenant := "foo"
+
+	httpReq := httptest.NewRequest("GET", api.PathMetricsQueryRange, nil)
+	httpReq = api.BuildQueryRangeRequest(httpReq, &tempopb.QueryRangeRequest{
+		Query: "{} | rate()",
+		Start: 1,
+		End:   uint64(10000 * time.Second),
+		Step:  uint64(1 * time.Second),
+	})
+
+	ctx := user.InjectOrgID(httpReq.Context(), tenant)
+	httpReq = httpReq.WithContext(ctx)
+
+	httpResp := httptest.NewRecorder()
+
+	f.MetricsQueryRangeHandler.ServeHTTP(httpResp, httpReq)
+
+	require.Equal(t, 200, httpResp.Code)
+
+	// for reasons I don't understand, this query turns into 408 jobs.
+	expectedResp := &tempopb.QueryRangeResponse{
+		Metrics: &tempopb.SearchMetrics{
+			CompletedJobs:   408,
+			InspectedTraces: 408,
+			InspectedBytes:  408,
+			TotalJobs:       408,
+			TotalBlocks:     2,
+			TotalBlockBytes: 419430400,
+		},
+		Series: []*tempopb.TimeSeries{
+			{
+				PromLabels: "foo",
+				Samples: []tempopb.Sample{
+					{
+						TimestampMs: 1,
+						Value:       408,
+					},
+					{
+						TimestampMs: 2,
+						Value:       816,
+					},
+				},
+			},
+		},
+	}
+
+	actualResp := &tempopb.QueryRangeResponse{}
+	err := jsonpb.Unmarshal(httpResp.Body, actualResp)
+	require.NoError(t, err)
+	require.Equal(t, expectedResp, actualResp)
+}
+
+func TestQueryRangeHandlerRespectsSamplingRate(t *testing.T) {
+	resp := &tempopb.QueryRangeResponse{
+		Metrics: &tempopb.SearchMetrics{
+			InspectedTraces: 1,
+			InspectedBytes:  1,
+		},
+		Series: []*tempopb.TimeSeries{
+			{
+				PromLabels: "foo",
+				Samples: []tempopb.Sample{
+					{
+						TimestampMs: 2,
+						Value:       2,
+					},
+					{
+						TimestampMs: 1,
+						Value:       1,
+					},
+				},
+			},
+		},
+	}
+
+	f := frontendWithSettings(t, &mockRoundTripper{
+		responseFn: func() proto.Message {
+			return resp
+		},
+	}, nil, nil, nil)
+	tenant := "foo"
+
+	httpReq := httptest.NewRequest("GET", api.PathMetricsQueryRange, nil)
+	httpReq = api.BuildQueryRangeRequest(httpReq, &tempopb.QueryRangeRequest{
+		Query: "{} | rate() with (sample=.2)",
+		Start: 1,
+		End:   uint64(10000 * time.Second),
+		Step:  uint64(1 * time.Second),
+	})
+
+	ctx := user.InjectOrgID(httpReq.Context(), tenant)
+	httpReq = httpReq.WithContext(ctx)
+
+	httpResp := httptest.NewRecorder()
+
+	f.MetricsQueryRangeHandler.ServeHTTP(httpResp, httpReq)
+
+	require.Equal(t, 200, httpResp.Code)
+
+	// for reasons I don't understand, this query turns into 408 jobs.
+	expectedResp := &tempopb.QueryRangeResponse{
+		Metrics: &tempopb.SearchMetrics{
+			CompletedJobs:   102,
+			InspectedTraces: 102,
+			InspectedBytes:  102,
+			TotalJobs:       102,
+			TotalBlocks:     2,
+			TotalBlockBytes: 419430400,
+		},
+		Series: []*tempopb.TimeSeries{
+			{
+				PromLabels: "foo",
+				Samples: []tempopb.Sample{
+					{
+						TimestampMs: 1,
+						Value:       510,
+					},
+					{
+						TimestampMs: 2,
+						Value:       1020,
+					},
+				},
+			},
+		},
+	}
+
+	actualResp := &tempopb.QueryRangeResponse{}
+	err := jsonpb.Unmarshal(httpResp.Body, actualResp)
+	require.NoError(t, err)
+	require.Equal(t, expectedResp, actualResp)
+}